Now running integration tests
This now runs integration tests. If they pass, it passes; if they fail,
it fails.

I am only running it with one integration test scene (one of the
included samples). In my tests, the Android device connected to the
Windows slave will not load the second scene and report results, but it
works fine with one scene (and it works find on the Mac).

Also, despite an option to create a folder with .xml files of the
results, I’ve had limited success getting the results back. When I run
it directly from the command line on the Mac with a relative folder it
does as expected, but it doesn’t do so when run via buildbot. With an
entire path, it does work, but I haven’t figured out how to do that in
a cross-platform way.